requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0
